Moved the gameState switch logic out into the gameEngine
