+++ src/pages/GameScreen.jsx
@@ -1,13 +1,13 @@
 import { useState, useEffect } from "react";
 import useHero from "../hooks/useHero";
 import useDungeon from "../hooks/useDungeon";
+import toggleGameState from "../utils/gameEngine";
 import CharacterInfo from "../components/CharacterInfo";
 import DungeonMap from "../components/DungeonMap";
 import OptionPrompts from "../components/OptionPrompts";
 import BattleScreen from "../components/BattleScreen";
 import BattleLog from "../components/BattleLog";
 import styles from "./GameScreen.module.css";
-
 
 // Images
 
@@ -17,7 +17,7 @@ export default function GameScreen() {
   const [gameState, setGameState] = useState("barracks");
   const [battleLog, setBattleLog] = useState([]);
   const [loot, setLoot] = useState([]);
-  const [eventResolved, setEventResolved] = useState(false);
+
   const [currentRoom, setCurrentRoom] = useState(null);
 
   const addToBattleLog = (message) => {
@@ -29,11 +29,30 @@ export default function GameScreen() {
       ...prevHero,
       inventory: [...prevHero.inventory, loot],
     }));
-  }
+  };
 
   const handleBattleResult = (updatedHero) => {
     updateHero(updatedHero);
-  }
+  };
+
+  const handleAction = (action, nextRoomId) => {
+    toggleGameState({
+      gameState,
+      action,
+      nextRoomId,
+      hero,
+      dungeon,
+      currentRoom,
+      setCurrentRoom,
+      setGameState,
+      setBattleLog,
+      addToBattleLog,
+      setLoot,
+      updateHero,
+      loot,
+      addLootToInventory,
+    });
+  };
 
   useEffect(() => {
     if (gameState === "barracks") {
@@ -41,111 +60,6 @@ export default function GameScreen() {
     }
   }, [gameState]);
 
-  const handleAction = (action, nextRoomId) => {
-    switch (gameState) {
-      case "barracks":
-        if (dungeon) {console.log(dungeon.name);}
-        if (action === "ADVENTURE") {
-          if (hero.health <= 1) {
-            alert("You need to sleep!");
-          } else {
-            setCurrentRoom(dungeon.rooms[0]);
-            setGameState("startAdventure");
-          }
-        }
-        if (action === "SLEEP") {
-          addToBattleLog("You are sleeping...");
-          setGameState("sleeping");
-        }
-        break;
-
-        case "startAdventure":
-        setGameState("enterRoom");
-        break;
-        
-        case "sleeping":
-        //Set characterState to full health, and make the player wait for a few seconds
-        updateHero({
-          ...hero,
-          health: hero.maxHealth,
-        });
-        addToBattleLog("You are fully healed!");
-        setGameState("barracks");
-        break;
-
-      case "enterRoom":
-        setGameState("encounter");
-        if (currentRoom.monsters.length) setGameState("battleChoice");
-        if (currentRoom.event) setGameState("resolveEvent");
-        break;
-
-      case "encounter":
-        setGameState("resolveEvent");
-        break;
-      case "resolveEvent":
-        setEventResolved(true);
-        addToBattleLog("Event resolved!");
-        setGameState("loot");
-        break;
-
-      case "battleChoice":
-        if (action === "FLEE") {
-          addToBattleLog("You fled the battle!");
-          setGameState("barracks");
-        }
-        if (action === "USE_ITEM") {
-          // implement item logic
-        }
-        if (action === "FIGHT") setGameState("autoBattle");
-        break;
-
-      case "autoBattle":
-        addToBattleLog("Resolving battle...");
-        
-        setGameState("battleOutcome");
-        break;
-
-      case "battleOutcome":
-        if (hero.health > 0) {
-          addToBattleLog("You won the battle!");
-          setLoot(currentRoom.monsters[0].loot.items[0]);
-          setGameState("loot");
-        } else {
-          addToBattleLog("You died!");
-          setGameState("barracks");
-        }
-        break;
-
-      case "loot":
-        // maybe the autobattler should set the loot?
-        // TODO: check if there is loot, otherwise just skip to next step
-        if (loot) {
-          addToBattleLog("You found loot:" + loot);
-          addLootToInventory(loot);
-          setLoot(null);
-        }
-        setGameState("continueOrHome");
-        break;
-
-      case "continueOrHome":
-        if (action === "CONTINUE") {
-          let nextRoom = dungeon.rooms.find((room) => room.id === nextRoomId);
-          if (nextRoom) {
-            setCurrentRoom(nextRoom);
-            setGameState("enterRoom");
-          } else {
-            // Handle case where nextRoomId is invalid
-            setBattleLog([...battleLog, "No more rooms to explore!"]);
-            setGameState("barracks");
-          }
-        }
-        if (action === "GO_HOME") setGameState("barracks");
-        break;
-
-      default:
-        break;
-    }
-  };
   return (
     <div className={styles.outer}>
       <div className={styles.content}>
